+++ stralloc/stralloc_starts.c
@@ -3,7 +3,7 @@
 #include "str.h"
 
 extern int stralloc_starts(stralloc *sa,const char *in) {
-  register unsigned long int len=str_len(in);
+  register size_t len=str_len(in);
   return (len<=sa->len && !byte_diff(sa->s,len,in));
 }
